requirements.txt
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tensorflow>=2.10.0
keras>=2.10.0
gradio>=4.0.0
Pillow>=9.0.0